+++ src/layout/WebHeader.vue
@@ -8,7 +8,7 @@
             <div class="row" slot="content-header" slot-scope="{closeMenu}">
                 <div class="col-6 collapse-brand">
                     <a href="https://demos.creative-tim.com/vue-argon-design-system/documentation/">
-                        <img src="img/brand/blue.png">
+                        <img src="@/img/gogo-logo.png">
                     </a>
                 </div>
                 <div class="col-6 collapse-close">
@@ -17,59 +17,16 @@
             </div>
 
             <ul class="navbar-nav navbar-nav-hover align-items-lg-center">
-                <li class="nav-item d-none d-lg-block ml-lg-4 item-muted">
-                    <router-link to="/homepage" class="">Home</router-link>
-                </li>
-                <li class="nav-item d-none d-lg-block ml-lg-4">
-                    <router-link to="/homepage" class="">About</router-link>
-                </li>
-                <li class="nav-item d-none d-lg-block ml-lg-4">
-                    <router-link to="/homepage" class="">Documentation</router-link>
-                </li>
-                <li class="nav-item d-none d-lg-block ml-lg-4">
-                    <router-link to="/homepage" class="">Resource</router-link>
-                </li>
-
-                <!-- <base-dropdown class="nav-item" menu-classes="dropdown-menu-xl">
-                    <a slot="title" href="#" class="nav-link" data-toggle="dropdown" role="button">
-                        <i class="ni ni-ui-04 d-lg-none"></i>
-                        <span class="nav-link-inner--text">Components</span>
-                    </a>
-                    <div class="dropdown-menu-inner">
-                        <a href="https://demos.creative-tim.com/vue-argon-design-system/documentation/"
-                           class="media d-flex align-items-center">
-                            <div class="icon icon-shape bg-gradient-primary rounded-circle text-white">
-                                <i class="ni ni-spaceship"></i>
-                            </div>
-                            <div class="media-body ml-3">
-                                <h6 class="heading text-primary mb-md-1">Getting started</h6>
-                                <p class="description d-none d-md-inline-block mb-0">Get started with Bootstrap, the
-                                    world's most popular framework for building responsive sites.</p>
-                            </div>
-                        </a>
-                        <a href="https://demos.creative-tim.com/vue-argon-design-system/documentation/"
-                           class="media d-flex align-items-center">
-                            <div class="icon icon-shape bg-gradient-warning rounded-circle text-white">
-                                <i class="ni ni-ui-04"></i>
-                            </div>
-                            <div class="media-body ml-3">
-                                <h5 class="heading text-warning mb-md-1">Components</h5>
-                                <p class="description d-none d-md-inline-block mb-0">Learn how to use Argon
-                                    compiling Scss, change brand colors and more.</p>
-                            </div>
-                        </a>
-                    </div>
-                </base-dropdown>
-                <base-dropdown tag="li" class="nav-item">
-                    <a slot="title" href="#" class="nav-link" data-toggle="dropdown" role="button">
-                        <i class="ni ni-collection d-lg-none"></i>
-                        <span class="nav-link-inner--text">Examples</span>
-                    </a>
-                    <router-link to="/component" class="dropdown-item">Component</router-link>
-                    <router-link to="/profile" class="dropdown-item">Profile</router-link>
-                    <router-link to="/login" class="dropdown-item">Login</router-link>
-                    <router-link to="/register" class="dropdown-item">Register</router-link>
-                </base-dropdown> -->
+                <a slot="title" href="#/homepage" class="nav-link" role="button">
+                    <span class="nav-link-inner--text">Home</span>
+                </a>
+                <a slot="title" href="#/about" class="nav-link" role="button">
+                    <span class="nav-link-inner--text">About</span>
+                </a>
+                <a slot="title" href="#/homepage" class="nav-link" role="button">
+                    <span class="nav-link-inner--text">Documentation</span>
+                </a>
+                
             </ul>
 
            <ul class="navbar-nav align-items-lg-center ml-lg-auto">