+++ CHANGELOG.md
@@ -1,11 +1,30 @@
 # Changelog
 
-##[1.1.0] 2019-02-12
-- Package updates
-- SSR fixes for checkboxes and radio
-- IE fixes
-- Best practices & accessibility improvements + pwa support
+## [0.1.0] 27-07-2021
+- Add homepage
+    - change header
+        - change nav bar format
+            - add gogo logo to header
+            - change nav bar item (home about documentation download)
+            - added Buy now with logo
+            - added contact's link
+    - add gogo logo with an animation (not interactive)
+    - add `Try This`
+        - use interactive component to show the example code
+    - add card components
+        - insert 3 supported platforms(ardruino , scratch , seeed)
+    - add interactive component
+        - embed the gogo code to the component
+    - change footer
+        - insert farewell
+        - change credit name to Logixed
 
-## [1.0.0] 2018-08-14
-Initial stable release
-### Original Release
+## [0.1.1] 09-08-2021
+- Use Full Bootstrap ('Vue.use(Bootstrap)')
+- Responsive UI
+- Fix card components
+    - add content to the supported platforms (e.g. Logo)
+- Fix footer format to align at center
+- Add FAQ at the bottom of the body by using bootstrap's tabs
+    - remains content
+- Fix navbar when it's become mobile (Hamburger at top right)